requirements.txt
loguru==0.5.3
matplotlib==3.3.4
networkx==2.5.1
numpy
pandas==1.1.5
Pillow==8.2.0
PyYAML==5.4.1
tensorboard==2.5.0
torch==1.9.0
yapf==0.31.0
